pyproject.toml
[project]
name = "gaatha-salesforce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.125.0",
    "granian>=2.6.0",
    "psycopg[binary]>=3.3.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sentry-sdk>=2.48.0",
    "structlog>=25.5.0",
    "uuid-utils>=0.12.0",
    "uvloop>=0.22.1",
    "yoyo-migrations>=9.0.0",
]

[dependency-groups]
dev = [
    "fastapi[standard]>=0.125.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "testcontainers[postgres]>=4.13.3",
]
